type promotion causing signed/unsigned comparison warnings

+++ src/lte/model/lte-ue-mac.cc
@@ -389,7 +389,7 @@ LteUeMac::DoReceiveIdealControlMessage (Ptr<IdealControlMessage> msg)
             {
               NS_LOG_FUNCTION (this << "\t" << dci.m_tbSize / m_macSapUserMap.size () << " bytes to LC " << (uint16_t)(*it).first << " queue " << (*itBsr).second);
               (*it).second->NotifyTxOpportunity (dci.m_tbSize / activeLcs, 0);
-              if ((*itBsr).second >=  dci.m_tbSize / activeLcs)
+              if ((*itBsr).second >=  static_cast<uint64_t> (dci.m_tbSize / activeLcs))
                 {
                   (*itBsr).second -= dci.m_tbSize / activeLcs;
                 }
